api/requirements.txt
flask==2.3.2
pyjwt==2.8.0
